GStreamerSharp: fix audio&video playback in 32bits (bgo#750041)
The playback was broken on 32 bits because of a bug in gstreamer-sharp
where the handles always were 64 bits long even on 32 bits. This is
fixed in master [1] (which we already depend on, since [2]).

This fix updates VideoManager to account for the changes (and this
does not only fix video, since playing audio apparently also calls
WindowRealize).

[1] http://cgit.freedesktop.org/gstreamer/gstreamer-sharp/commit/?id=9556b05947a43cbeea57113f427d6bde5246c77f
[2] https://git.gnome.org/browse/banshee/commit/?id=e35092655886a6cce314d2310404f977233588a4

+++ src/Backends/Banshee.GStreamerSharp/Banshee.GStreamerSharp/VideoManager.cs
@@ -52,7 +52,7 @@ namespace Banshee.GStreamerSharp
         Element playbin;
         VideoDisplayContextType video_display_context_type;
         IntPtr video_window;
-        ulong? video_window_xid;
+        IntPtr? video_window_xid;
         VideoOverlayAdapter xoverlay;
         object video_mutex = new object ();
 
@@ -295,10 +295,10 @@ namespace Banshee.GStreamerSharp
         public void WindowRealize (IntPtr window)
         {
             if (PlatformDetection.IsWindows) {
-                video_window_xid = (ulong)gdk_win32_drawable_get_handle (window);
+                video_window_xid = gdk_win32_drawable_get_handle (window);
             } else if (PlatformDetection.IsUnix) {
                     //FIXME: we should maybe stop relying on x11 http://gstreamer.freedesktop.org/data/doc/gstreamer/head/gst-plugins-base-libs/html/gst-plugins-base-libs-gstvideooverlay.html#GstVideoOverlay
-                video_window_xid = (ulong)gdk_x11_window_get_xid (window);
+                video_window_xid = gdk_x11_window_get_xid (window);
             }
         }
